python-api/requirements.txt
wheel
flask
flask_sslify
flask_cors
flask_login
flask_migrate
python-dotenv
gunicorn
requests
ibm-watson
Werkzeug